(update) fix when rendering weight alternatif.

+++ database/migrations/2023_06_17_160402_create_kmeans_data_table.php
@@ -14,16 +14,16 @@ return new class extends Migration
         Schema::create('kmeans_data', function (Blueprint $table) {
             $table->id();
             $table->string('nama_pemilik');
-            $table->decimal('jumlah_pekerja');
-            $table->decimal('jenis_produksi');
-            $table->decimal('kapasitas_produksi');
-            $table->decimal('harga_satuan', 10, 2);
-            $table->decimal('nilai_produksi', 10, 2);
-            $table->decimal('nilai_investasi', 10, 2);
-            $table->decimal('umur');
-            $table->decimal('pendidikan');
-            $table->decimal('surat_izin');
-            $table->decimal('motif');
+            $table->float('jumlah_pekerja');
+            $table->float('jenis_produksi');
+            $table->float('kapasitas_produksi');
+            $table->float('harga_satuan');
+            $table->float('nilai_produksi');
+            $table->float('nilai_investasi');
+            $table->float('umur');
+            $table->float('pendidikan');
+            $table->float('surat_izin');
+            $table->float('motif');
             $table->timestamps();
         });
     }
